test: answer 리포지토리 테스트 수정
